Refactor tests
Separate tests into dedicated files
